Enable headless builds using Maven Tycho.
Further changes:

- Remove org.beanaire.ui.e4.notification.lib/sac-1.3.jar lib.
  Include org.w3c.css.sac bundle instead (as a dependency).

- Remove Eclipse specific files from repo and add gitignore file.

- Include CSS file in binary build.
  (org.beanaire.ui.e4.notification.lib/build.properties).

- Include README.md.
