+++ .gitignore
@@ -1,0 +1,7 @@
+# Eclipse files
+.project
+.classpath
+
+# Eclipse folders
+.settings
+target
