+++ org.beanaire.ui.e4.notification.demo.target/org.beanaire.ui.e4.notification.demo.target.target
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<?pde version="3.8"?>
+
+<target name="org.beanaire.ui.e4.notification.demo">
+
+	<locations>
+
+		<!-- ECLIPSE RELEASES REPO -->
+		<location includeAllPlatforms="false" includeConfigurePhase="true" includeMode="planner" includeSource="true" type="InstallableUnit">
+			<repository location="http://download.eclipse.org/releases/mars"/>
+			<unit id="org.eclipse.rcp.feature.group" version="0.0.0"/>
+			<unit id="org.eclipse.equinox.executable.feature.group" version="0.0.0"/>
+		</location>
+
+	</locations>
+
+</target>
