+++ org.beanaire.ui.e4.notification.demo.target/pom.xml
@@ -1,0 +1,16 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>org.beanaire.ui.e4.notification.demo.target</artifactId>
+
+	<packaging>eclipse-target-definition</packaging>
+
+	<parent>
+		<groupId>org.beanaire.ui.e4.notification</groupId>
+		<artifactId>org.beanaire.ui.e4.notification.parent</artifactId>
+		<version>1.1.0</version>
+		<relativePath>../org.beanaire.ui.e4.notification.parent/pom.xml</relativePath>
+	</parent>
+</project>
