+++ org.beanaire.ui.e4.notification.demo/META-INF/MANIFEST.MF
@@ -1,8 +1,8 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
-Bundle-Name: Demo Notification
+Bundle-Name: Demo Notification Plugin
 Bundle-SymbolicName: org.beanaire.ui.e4.notification.demo;singleton:=true
-Bundle-Version: 1.0.0.qualifier
+Bundle-Version: 1.0.1.qualifier
 Bundle-Vendor: www.beanaire.fr
 Require-Bundle: javax.inject;bundle-version="1.0.0",
  org.eclipse.core.runtime;bundle-version="3.9.0",
@@ -14,6 +14,6 @@ Require-Bundle: javax.inject;bundle-version="1.0.0",
  org.eclipse.e4.core.di;bundle-version="1.3.0",
  org.eclipse.e4.ui.di;bundle-version="1.0.0",
  org.eclipse.e4.core.contexts;bundle-version="1.3.0",
- org.beanaire.ui.e4.notification
+ org.beanaire.ui.e4.notification.lib
 Bundle-RequiredExecutionEnvironment: JavaSE-1.6
 Import-Package: javax.annotation;version="1.0.0"
