+++ org.beanaire.ui.e4.notification.demo/plugin.xml
@@ -1,18 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <?eclipse version="3.4"?>
 <plugin>
-
    <extension
          id="product"
          point="org.eclipse.core.runtime.products">
       <product
             application="org.eclipse.e4.ui.workbench.swt.E4Application"
-            name="org.beanaire.ui.e4.notification.demo">
+            name="Demo Notification Product">
          <property
                name="appName"
-               value="org.beanaire.ui.e4.notification.demo">
+               value="Demo Notification Product">
          </property>
       </product>
    </extension>
+
 
 </plugin>
