+++ org.beanaire.ui.e4.notification.lib.parent/pom.xml
@@ -1,0 +1,44 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>org.beanaire.ui.e4.notification.lib.parent</artifactId>
+
+	<packaging>pom</packaging>
+
+	<build>
+		<plugins>
+
+			<plugin>
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>target-platform-configuration</artifactId>
+				<version>${tycho-version}</version>
+				<configuration>
+					<target>
+						<artifact>
+							<groupId>org.beanaire.ui.e4.notification</groupId>
+							<!-- This plugin expects exactly one file (<artifactId>.target) in 
+								the project's base directory. -->
+							<artifactId>org.beanaire.ui.e4.notification.lib.target</artifactId>
+							<version>1.0.1-SNAPSHOT</version>
+						</artifact>
+					</target>
+				</configuration>
+			</plugin>
+
+		</plugins>
+	</build>
+
+	<modules>
+		<module>../org.beanaire.ui.e4.notification.lib.target</module>
+		<module>../org.beanaire.ui.e4.notification.lib</module>
+	</modules>
+
+	<parent>
+		<groupId>org.beanaire.ui.e4.notification</groupId>
+		<artifactId>org.beanaire.ui.e4.notification.parent</artifactId>
+		<relativePath>../org.beanaire.ui.e4.notification.parent</relativePath>
+		<version>1.0.1-SNAPSHOT</version>
+	</parent>
+</project>
