+++ org.beanaire.ui.e4.notification.lib.target/org.beanaire.ui.e4.notification.lib.target.target
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="UTF-8" standalone="no"?>
+<?pde version="3.8"?>
+
+<target name="org.beanaire.ui.e4.notification.lib">
+
+	<locations>
+
+		<!-- ECLIPSE RELEASES REPO -->
+		<location includeAllPlatforms="false" includeConfigurePhase="true" includeMode="planner" includeSource="true" type="InstallableUnit">
+			<repository location="http://download.eclipse.org/releases/mars"/>
+			<unit id="org.eclipse.rcp.feature.group" version="0.0.0"/>
+		</location>
+
+	</locations>
+
+</target>
