+++ org.beanaire.ui.e4.notification.lib/META-INF/MANIFEST.MF
@@ -1,0 +1,17 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Notification UI System for E4
+Bundle-SymbolicName: org.beanaire.ui.e4.notification.lib;singleton:=true
+Bundle-Version: 1.0.1.qualifier
+Import-Package: org.eclipse.core.runtime;version="3.4.0",
+ org.osgi.framework;version="1.3.0"
+Bundle-ActivationPolicy: lazy
+Require-Bundle: org.eclipse.swt,
+ org.eclipse.e4.ui.css.core,
+ org.eclipse.e4.ui.css.swt,
+ org.eclipse.jface,
+ org.w3c.css.sac
+Export-Package: org.beanaire.ui.e4.notification
+Bundle-RequiredExecutionEnvironment: JavaSE-1.6
+Bundle-ClassPath: sac-1.3.jar,
+ .
