+++ org.beanaire.ui.e4.notification.lib/build.properties
@@ -1,0 +1,6 @@
+output.. = target/classes/
+bin.includes = META-INF/,\
+               .,\
+               images/,\
+               css/
+source.. = src/main/java/
