+++ org.beanaire.ui.e4.notification.lib/css/notify.css
@@ -1,0 +1,8 @@
+CLabel.title {
+	font : Arial bold 12px;
+}
+
+Label.message {
+	font : Arial 10px;
+}
+
