+++ org.beanaire.ui.e4.notification.lib/images/fileclose_16x16.png
