+++ org.beanaire.ui.e4.notification.lib/pom.xml
@@ -1,0 +1,17 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<artifactId>org.beanaire.ui.e4.notification.lib</artifactId>
+
+	<parent>
+		<groupId>org.beanaire.ui.e4.notification</groupId>
+		<artifactId>org.beanaire.ui.e4.notification.lib.parent</artifactId>
+		<version>1.0.1-SNAPSHOT</version>
+		<relativePath>../org.beanaire.ui.e4.notification.lib.parent/pom.xml</relativePath>
+	</parent>
+
+	<packaging>eclipse-plugin</packaging>
+
+</project>
