+++ org.beanaire.ui.e4.notification.parent/pom.xml
@@ -1,0 +1,38 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<groupId>org.beanaire.ui.e4.notification</groupId>
+	<artifactId>org.beanaire.ui.e4.notification.parent</artifactId>
+	<version>1.0.1-SNAPSHOT</version>
+
+	<packaging>pom</packaging>
+
+	<properties>
+		<!-- See http://mvnrepository.com/artifact/org.eclipse.tycho/tycho-maven-plugin/ 
+			for latest version. -->
+		<tycho-version>0.24.0</tycho-version>
+	</properties>
+
+	<build>
+		<plugins>
+
+			<plugin>
+				<!-- enable tycho build extension -->
+				<groupId>org.eclipse.tycho</groupId>
+				<artifactId>tycho-maven-plugin</artifactId>
+				<version>${tycho-version}</version>
+				<!-- The Tycho plugin needs to be defined as an extension. Otherwise 
+					it does not contribute to the packaging type. -->
+				<extensions>true</extensions>
+			</plugin>
+
+		</plugins>
+	</build>
+
+	<modules>
+		<module>../org.beanaire.ui.e4.notification.demo.parent</module>
+		<module>../org.beanaire.ui.e4.notification.lib.parent</module>
+	</modules>
+</project>
