--- org.beanaire.ui.e4.notification/.settings/org.eclipse.pde.core.prefs
@@ -1,3 +1,0 @@
-eclipse.preferences.version=1
-pluginProject.extensions=false
-resolve.requirebundle=false
