--- org.beanaire.ui.e4.notification/build.properties
@@ -1,6 +1,0 @@
-output.. = target/classes/
-bin.includes = META-INF/,\
-               .,\
-               sac-1.3.jar,\
-               images/
-source.. = src/main/java/
