--- org.beanaire.ui.e4.notification/css/notify.css
@@ -1,8 +1,0 @@
-CLabel.title {
-	font : Arial bold 12px;
-}
-
-Label.message {
-	font : Arial 10px;
-}
-
